Extract reusable label creating logic to own component
